[WFCORE-6659] Remove ServiceController.getValue use from embedded server/hc and bootable jar server start
Note that this changes the behavior of start for these classes by blocking until the ModelControllerClientFactory becomes available.
Currently they just assume it is available future return by the bootstrap is, but that is an unsafe assumption,
and one that can only work if the deprecated MSC Service interface is used with a service whose value is passed to the c'tor.